WAL-91 feat: DynamoDB services

+++ test/services/account.js
@@ -1,37 +1,42 @@
+const { USE_DYNAMODB } = require('../../features');
 const AccountService = require('../../services/account');
 const chai = require('../chai');
 const { deleteAllRows } = require('../db');
+const initLocalDynamo = require('../local_dynamo');
+const { generateAccountId } = require('../utils');
 
 const { expect } = chai;
 
-function* generateAccountId() {
-    let accountSuffix = 1;
-    while (true) {
-        yield `account${accountSuffix++}.near`;
-    }
-}
-
-const accountGenerator = generateAccountId();
-function getAccountId() {
-    return accountGenerator.next().value;
-}
+const accountService = new AccountService();
 
 describe('AccountService', function () {
+    let terminateLocalDynamo;
     before(async function() {
-        await deleteAllRows();
+        if (USE_DYNAMODB) {
+            this.timeout(10000);
+            ({ terminateLocalDynamo } = await initLocalDynamo());
+        } else {
+            await deleteAllRows();
+        }
+    });
+
+    after(async function() {
+        if (USE_DYNAMODB) {
+            await terminateLocalDynamo();
+        }
     });
 
     describe('createAccount', function () {
         it('accounts are created with default parameters', async function () {
-            const accountId = getAccountId();
-            const account = await AccountService.createAccount(accountId);
+            const accountId = generateAccountId();
+            const account = await accountService.createAccount(accountId);
             expect(account).property('accountId', accountId);
             expect(account).property('fundedAccountNeedsDeposit', false);
         });
 
         it('accounts are correctly created with fundedAccountNeedsDeposit', async function () {
-            const accountId = getAccountId();
-            const account = await AccountService.createAccount(accountId, { fundedAccountNeedsDeposit: true });
+            const accountId = generateAccountId();
+            const account = await accountService.createAccount(accountId, { fundedAccountNeedsDeposit: true });
             expect(account).property('accountId', accountId);
             expect(account).property('fundedAccountNeedsDeposit', true);
         });
@@ -39,41 +44,41 @@ describe('AccountService', function () {
 
     describe('deleteAccount', function () {
         it('the account with matching accountId is deleted', async function () {
-            const accountId = getAccountId();
-            let account = await AccountService.createAccount(accountId);
+            const accountId = generateAccountId();
+            let account = await accountService.createAccount(accountId);
             expect(account).property('accountId', accountId);
 
-            await AccountService.deleteAccount(accountId);
-            account = await AccountService.getAccount(accountId);
+            await accountService.deleteAccount(accountId);
+            account = await accountService.getAccount(accountId);
             expect(account).null;
         });
     });
 
     describe('getAccount', function () {
         it('the account with matching accountId is returned', async function () {
-            const accountId = getAccountId();
-            let account = await AccountService.createAccount(accountId);
+            const accountId = generateAccountId();
+            let account = await accountService.createAccount(accountId);
             expect(account).property('accountId', accountId);
 
-            account = await AccountService.getAccount(accountId);
+            account = await accountService.getAccount(accountId);
             expect(account).property('accountId', accountId);
         });
 
         it('returns null for non-existent accounts', async function () {
             const accountId = 'nonexistent.near';
-            const account = await AccountService.getAccount(accountId);
+            const account = await accountService.getAccount(accountId);
             expect(account).null;
         });
     });
 
     describe('setAccountRequiresDeposit', function () {
         it('sets the fundedAccountNeedsDeposit flag', async function () {
-            const accountId = getAccountId();
-            let account = await AccountService.createAccount(accountId);
+            const accountId = generateAccountId();
+            let account = await accountService.createAccount(accountId);
             expect(account).property('fundedAccountNeedsDeposit', false);
 
-            await AccountService.setAccountRequiresDeposit(accountId, true);
-            account = await AccountService.getAccount(accountId);
+            await accountService.setAccountRequiresDeposit(accountId, true);
+            account = await accountService.getAccount(accountId);
             expect(account).property('fundedAccountNeedsDeposit', true);
         });
     });
