+++ test/services/identity_verification_method.js
@@ -3,49 +3,82 @@ const IdentityVerificationMethodService = require('../../services/identity_verif
 const IdentityVerificationMethodSequelize = require('../../services/sequelize/identity_verification_method');
 const chai = require('../chai');
 const { deleteAllRows } = require('../db');
+const { USE_DYNAMODB } = require('../../features');
+const initLocalDynamo = require('../local_dynamo');
+const { generateEmailAddress } = require('../utils');
 
 const { IDENTITY_VERIFICATION_METHOD_KINDS } = Constants;
 const { expect } = chai;
 
-const EMAIL = 'test@near.org';
-const SECURITY_CODE = 123456;
+const SECURITY_CODE = '123456';
+
+const identityVerificationMethodService = new IdentityVerificationMethodService();
 
 describe('IdentityVerificationMethodService', function () {
+    let terminateLocalDynamo;
     before(async function() {
-        await deleteAllRows();
+        if (USE_DYNAMODB) {
+            this.timeout(10000);
+            ({ terminateLocalDynamo } = await initLocalDynamo());
+        } else {
+            await deleteAllRows();
+        }
+    });
+
+    after(async function() {
+        if (USE_DYNAMODB) {
+            await terminateLocalDynamo();
+        }
     });
 
     describe('claimIdentityVerificationMethod', function () {
         beforeEach(async function () {
-            await deleteAllRows();
+            if (!USE_DYNAMODB) {
+                await deleteAllRows();
+            }
         });
 
         it('updates an unclaimed identity verification method to be claimed', async function () {
             const params = {
-                identityKey: EMAIL,
+                identityKey: generateEmailAddress(),
                 kind: IDENTITY_VERIFICATION_METHOD_KINDS.EMAIL,
             };
 
-            await IdentityVerificationMethodSequelize.createIdentityVerificationMethod_internal({
-                ...params,
-                securityCode: SECURITY_CODE,
-            });
-            await IdentityVerificationMethodService.claimIdentityVerificationMethod(params);
+            if (USE_DYNAMODB) {
+                await identityVerificationMethodService.recoverIdentity({
+                    ...params,
+                    securityCode: SECURITY_CODE,
+                });
+            } else {
+                await IdentityVerificationMethodSequelize.createIdentityVerificationMethod_internal({
+                    ...params,
+                    securityCode: SECURITY_CODE,
+                });
+            }
 
-            const identityVerificationMethod = await IdentityVerificationMethodService.getIdentityVerificationMethod(params);
-            expect(identityVerificationMethod).property('securityCode', null);
+            await identityVerificationMethodService.claimIdentityVerificationMethod(params);
+
+            const identityVerificationMethod = await identityVerificationMethodService.getIdentityVerificationMethod(params);
+            if (USE_DYNAMODB) {
+                expect(identityVerificationMethod).not.have.property('securityCode');
+            } else {
+                expect(identityVerificationMethod).property('securityCode', null);
+            }
+
             expect(identityVerificationMethod).property('claimed', true);
         });
     });
 
     describe('getIdentityVerificationMethod', function () {
         beforeEach(async function () {
-            await deleteAllRows();
+            if (!USE_DYNAMODB) {
+                await deleteAllRows();
+            }
         });
 
         it('returns null for non-existent records', async function () {
-            const identityVerificationMethod = await IdentityVerificationMethodService.getIdentityVerificationMethod({
-                identityKey: 'fake@gmail.com',
+            const identityVerificationMethod = await identityVerificationMethodService.getIdentityVerificationMethod({
+                identityKey: generateEmailAddress(),
                 kind: IDENTITY_VERIFICATION_METHOD_KINDS.EMAIL,
             });
 
@@ -54,63 +87,80 @@ describe('IdentityVerificationMethodService', function () {
 
         it('gets the identity verification method based on identityKey and kind', async function () {
             const params = {
-                identityKey: EMAIL,
+                identityKey: generateEmailAddress(),
                 kind: IDENTITY_VERIFICATION_METHOD_KINDS.EMAIL,
             };
 
-            await IdentityVerificationMethodSequelize.createIdentityVerificationMethod_internal({
-                ...params,
-                securityCode: SECURITY_CODE,
-            });
+            if (USE_DYNAMODB) {
+                await identityVerificationMethodService.recoverIdentity({
+                    ...params,
+                    securityCode: SECURITY_CODE,
+                });
+            } else {
+                await IdentityVerificationMethodSequelize.createIdentityVerificationMethod_internal({
+                    ...params,
+                    securityCode: SECURITY_CODE,
+                });
+            }
 
-            const identityVerificationMethod = await IdentityVerificationMethodService.getIdentityVerificationMethod(params);
-            expect(identityVerificationMethod).property('identityKey', EMAIL);
-            expect(identityVerificationMethod).property('kind', IDENTITY_VERIFICATION_METHOD_KINDS.EMAIL);
+            const identityVerificationMethod = await identityVerificationMethodService.getIdentityVerificationMethod(params);
+            expect(identityVerificationMethod).property('identityKey', params.identityKey);
+            expect(identityVerificationMethod).property('kind', params.kind);
         });
     });
 
     describe('getUniqueEmail', function () {
         it('returns empty string for email addresses without @', async function () {
-            expect(IdentityVerificationMethodService.getUniqueEmail('xyz')).equal('');
+            expect(identityVerificationMethodService.getUniqueEmail('xyz')).equal('');
         });
 
         it('returns the same value for googlemail addresses', async function () {
-            const gmailAddress = IdentityVerificationMethodService.getUniqueEmail('test@gmail.com');
-            const googleMailAddress = IdentityVerificationMethodService.getUniqueEmail('test@googlemail.com');
+            const gmailAddress = identityVerificationMethodService.getUniqueEmail('test@gmail.com');
+            const googleMailAddress = identityVerificationMethodService.getUniqueEmail('test@googlemail.com');
             expect(gmailAddress).equal(googleMailAddress);
         });
 
         it('considers email addresses with + characters to be equivalent', async function () {
-            const gmailAddress = IdentityVerificationMethodService.getUniqueEmail('test@gmail.com');
-            const subAddress = IdentityVerificationMethodService.getUniqueEmail('test+tessst@gmail.com');
+            const gmailAddress = identityVerificationMethodService.getUniqueEmail('test@gmail.com');
+            const subAddress = identityVerificationMethodService.getUniqueEmail('test+tessst@gmail.com');
             expect(gmailAddress).equal(subAddress);
         });
 
         it('ignores special characters in email username', async function () {
-            const gmailAddress = IdentityVerificationMethodService.getUniqueEmail('test@gmail.com');
-            const invalidCharAddress = IdentityVerificationMethodService.getUniqueEmail('t|(e)#s>t@gmail.com');
+            const gmailAddress = identityVerificationMethodService.getUniqueEmail('test@gmail.com');
+            const invalidCharAddress = identityVerificationMethodService.getUniqueEmail('t|(e)#s>t@gmail.com');
             expect(gmailAddress).equal(invalidCharAddress);
         });
     });
 
     describe('recoverIdentity', function () {
         beforeEach(async function () {
-            await deleteAllRows();
+            if (!USE_DYNAMODB) {
+                await deleteAllRows();
+            }
         });
 
         it('returns false for already-claimed identity verification methods', async function () {
             const params = {
-                identityKey: EMAIL,
+                identityKey: generateEmailAddress(),
                 kind: IDENTITY_VERIFICATION_METHOD_KINDS.EMAIL,
             };
 
-            await IdentityVerificationMethodSequelize.createIdentityVerificationMethod_internal({
-                ...params,
-                securityCode: SECURITY_CODE,
-            });
-            await IdentityVerificationMethodService.claimIdentityVerificationMethod(params);
+            if (USE_DYNAMODB) {
+                await identityVerificationMethodService.recoverIdentity({
+                    ...params,
+                    securityCode: SECURITY_CODE,
+                });
+            } else {
+                await IdentityVerificationMethodSequelize.createIdentityVerificationMethod_internal({
+                    ...params,
+                    securityCode: SECURITY_CODE,
+                });
+            }
 
-            const isRecovered = await IdentityVerificationMethodService.recoverIdentity({
+            await identityVerificationMethodService.claimIdentityVerificationMethod(params);
+
+            const isRecovered = await identityVerificationMethodService.recoverIdentity({
                 ...params,
                 securityCode: SECURITY_CODE,
             });
@@ -120,16 +170,21 @@ describe('IdentityVerificationMethodService', function () {
 
         it('updates unclaimed identity verification methods', async function () {
             const params = {
-                identityKey: EMAIL,
+                identityKey: generateEmailAddress(),
                 kind: IDENTITY_VERIFICATION_METHOD_KINDS.EMAIL,
             };
 
-            await IdentityVerificationMethodSequelize.createIdentityVerificationMethod_internal(params);
-            const isRecovered = await IdentityVerificationMethodService.recoverIdentity({
+            if (USE_DYNAMODB) {
+                await identityVerificationMethodService.recoverIdentity(params);
+            } else {
+                await IdentityVerificationMethodSequelize.createIdentityVerificationMethod_internal(params);
+            }
+
+            const isRecovered = await identityVerificationMethodService.recoverIdentity({
                 ...params,
                 securityCode: SECURITY_CODE,
             });
-            const identityVerificationMethod = await IdentityVerificationMethodService.getIdentityVerificationMethod(params);
+            const identityVerificationMethod = await identityVerificationMethodService.getIdentityVerificationMethod(params);
 
             expect(isRecovered).true;
             expect(identityVerificationMethod).property('securityCode', SECURITY_CODE.toString());
@@ -137,12 +192,17 @@ describe('IdentityVerificationMethodService', function () {
 
         it('returns false when the identityKey exists for a different value for kind', async function () {
             const params = {
-                identityKey: EMAIL,
+                identityKey: generateEmailAddress(),
                 kind: IDENTITY_VERIFICATION_METHOD_KINDS.EMAIL,
             };
 
-            await IdentityVerificationMethodSequelize.createIdentityVerificationMethod_internal(params);
-            const isRecovered = await IdentityVerificationMethodService.recoverIdentity({
+            if (USE_DYNAMODB) {
+                await identityVerificationMethodService.recoverIdentity(params);
+            } else {
+                await IdentityVerificationMethodSequelize.createIdentityVerificationMethod_internal(params);
+            }
+
+            const isRecovered = await identityVerificationMethodService.recoverIdentity({
                 ...params,
                 kind: IDENTITY_VERIFICATION_METHOD_KINDS.PHONE,
                 securityCode: SECURITY_CODE,
@@ -157,11 +217,20 @@ describe('IdentityVerificationMethodService', function () {
                 kind: IDENTITY_VERIFICATION_METHOD_KINDS.EMAIL,
             };
 
-            await IdentityVerificationMethodSequelize.createIdentityVerificationMethod_internal({
-                ...params,
-                identityKey: 'test+test@gmail.com',
-            });
-            const isRecovered = await IdentityVerificationMethodService.recoverIdentity({
+
+            if (USE_DYNAMODB) {
+                await identityVerificationMethodService.recoverIdentity({
+                    ...params,
+                    identityKey: 'test+test@gmail.com',
+                });
+            } else {
+                await IdentityVerificationMethodSequelize.createIdentityVerificationMethod_internal({
+                    ...params,
+                    identityKey: 'test+test@gmail.com',
+                });
+            }
+
+            const isRecovered = await identityVerificationMethodService.recoverIdentity({
                 ...params,
                 kind: IDENTITY_VERIFICATION_METHOD_KINDS.EMAIL,
                 securityCode: SECURITY_CODE,
